split ActivityFormScreen into multiple files
