+++ src/screens/index.js
@@ -3,7 +3,7 @@ export { default as WeekScreen } from './WeekScreen'
 export { default as ActivityDetailScreen} from './ActivityDetailScreen/ActivityDetailScreen'
 export { default as GoalsScreen } from './GoalsScreen'
 export { default as GoalScreen} from './GoalScreen'
-export { default as ActivityFormScreen } from './ActivityFormScreen/ActivityFormScreen'
+export { default as ActivityFormScreen } from './ActivityFormScreen'
 export { default as GoalFormScreen } from './GoalFormScreen'
 export { default as CalendarScreen } from './CalendarScreen'
 export { default as SettingsScreen } from './SettingsScreen'
